Split `AttributesTreeValue` into own file
